Safari-611.1.10/LayoutTests/crypto/subtle/ec-import-jwk-key-export-jwk-key-private-expected.txt

Test importing a jwk private ECDH key and then export it in jwk format

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Importing a key...
PASS privateKey.kty is jwkKey.kty
PASS privateKey.crv is jwkKey.crv
PASS privateKey.x is jwkKey.x
PASS privateKey.y is jwkKey.y
PASS privateKey.d is jwkKey.d
PASS privateKey.key_ops is jwkKey.key_ops
PASS privateKey.ext is jwkKey.ext
PASS successfullyParsed is true

TEST COMPLETE
